Moved data/ and jsonld/ out of src/.
