+++ jsonld/convert2ttl.sh
