+++ jsonld/scripts/doi2bibjson.js
