+++ src/pages/projects.js
@@ -15,9 +15,9 @@ import Layout from "../components/layout"
 
 import css from "./projects.module.css"
 
-import projectsData from "../jsonld/projects.json"
-import publicationsData from "../jsonld/publications.json"
-import citationsData from "../jsonld/citations.json"
+import projectsData from "../../jsonld/projects.json"
+import publicationsData from "../../jsonld/publications.json"
+import citationsData from "../../jsonld/citations.json"
 
 export default function TimelinePage() {
   const refs = new ReferenceManager();
